Added Test to classes Manage for DictionarySentenses,Levels,Stats and ReaderCSVTest was started

+++ project/src/main/java/pl/competencyproject/model/DAO/interfaces/ManagingStats.java
@@ -9,7 +9,7 @@ public interface ManagingStats {
 
     public Stat getStat(int IdStat);
 
-    public void updateStat(int IdStat, int valueProgress);
+    public void updateStat(int IdStat, double valueProgress);
 
     public void deleteStat(int IdStat);
